% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vcf_files.R
\name{mc.loadVcf}
\alias{mc.loadVcf}
\title{Read one VCF file and create a dataframe}
\usage{
mc.loadVcf(vcf.file, ref.genome)
}
\arguments{
\item{vcf.file}{Name of the VCF file}

\item{ref.genome}{Name of the reference genome to use}
}
\value{
A dataframe with the VCF infos
}
\description{
Read one VCF file and create a dataframe
}

