% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vcf_files.R
\name{mc.loadVcfs}
\alias{mc.loadVcfs}
\title{Read VCF files as a list of data frames}
\usage{
mc.loadVcfs(vcf.files, sample.names, ref.genome)
}
\arguments{
\item{vcf.files}{A character vector of vcf file names}

\item{sample.names}{A character vector of sample names}

\item{ref.genome}{Name of the loaded reference genome}
}
\value{
A list of dataframes, one for each vcf file
}
\description{
Read VCF files as a list of data frames
}

