% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_context.R
\name{mc.mutContext}
\alias{mc.mutContext}
\title{Add context bases to mutations}
\usage{
mc.mutContext(mutnet.obj, ref.genome)
}
\arguments{
\item{mutnet.obj}{A dataframe of mutations}

\item{ref.genome}{Name of the loaded reference genome}
}
\value{
The dataframe with context bases added
}
\description{
Add context bases to mutations
}

