% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_spectrum.R
\name{mc.mutSpectrum}
\alias{mc.mutSpectrum}
\title{Make mutations table}
\usage{
mc.mutSpectrum(mutnet.obj, col.ref = "ref.allele", col.alt = "alt.allele",
  nuc = c("A", "C", "G", "T"), symetrize = T)
}
\arguments{
\item{mutnet.obj}{A dataframe of mutation info}

\item{col.ref}{Name of the column with the reference alleles}

\item{col.alt}{Name of the column with the alt alleles}

\item{nuc}{Array of used nucleotides}

\item{symetrize}{Boolean to symetrize reversable mutations or not}
}
\value{
A table with the counts and proportions of each mutation type
}
\description{
Make mutations table
}

