% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_context.R
\name{mc.plotContext}
\alias{mc.plotContext}
\title{Plot histogram of the 96 possible mutations}
\usage{
mc.plotContext(mutnet, base.ref.col = "ref.allele",
  base.alt.col = "alt.allele", base.3p.col = "base.3p",
  base.5p.col = "base.5p", trinuc.pc, symetrize = c(TRUE, FALSE)[1])
}
\arguments{
\item{mutnet}{A list of dataframe with mutations infos for each sample}

\item{base.ref.col}{Name of the column with the reference allele in dataframe}

\item{base.alt.col}{Name of the column with the alternative allele in dataframe}

\item{base.3p.col}{Name of the column with the base after the mutation in dataframe}

\item{base.5p.col}{Name of the column with the base before the mutation in dataframe}

\item{trinuc.pc}{Array of the proportion of each trinucleotide in the genome (for normalisation)}

\item{symetrize}{Boolean to symetrize reversable mutations or not}
}
\value{
A ggplot object to plot
}
\description{
Plot histogram of the 96 possible mutations
}

