% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rain.R
\name{mc.plotRain}
\alias{mc.plotRain}
\title{Plot mutation rainfalls}
\usage{
mc.plotRain(mutrain, chrom.to.plot = NULL, color.col = "mut")
}
\arguments{
\item{mutrain}{Table of distance between successive mutations in genomic order}

\item{chrom.to.plot}{List of the chromosomes to include}

\item{color.col}{Column of the table to use for colors}
}
\value{
A ggplot2 object to plot
}
\description{
Plot mutation rainfalls
}

