% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{mc.plotSpectrum}
\alias{mc.plotSpectrum}
\title{Plot mutation spectrum}
\usage{
mc.plotSpectrum(mut.sum, yaxis = c("count", "prop")[1], print.num = c(TRUE,
  FALSE)[1])
}
\arguments{
\item{mut.sum}{Table of the counts and proportions of each mutation type}

\item{yaxis}{Variable to plot: counts or proportions}

\item{print.num}{Boolean to print values on the plot or not}
}
\value{
A ggplot object to plot
}
\description{
Plot mutation spectrum
}

