% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spider.R
\name{mc.plotSpider}
\alias{mc.plotSpider}
\title{Plot spider spectrum of mutations}
\usage{
mc.plotSpider(mut.sum)
}
\arguments{
\item{mut.sum}{Table of the counts and proportions of each mutation type}
}
\value{
A ggplot object to plot
}
\description{
Plot spider spectrum of mutations
}

