% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_context.R
\name{mc.sumContext}
\alias{mc.sumContext}
\title{Compute frequencies and proportions of mutations with context}
\usage{
mc.sumContext(mutnet.obj, base.ref.col = "ref.allele",
  base.alt.col = "alt.allele", base.3p.col = "base.3p",
  base.5p.col = "base.5p", trinuc.pc, symetrize = c(TRUE, FALSE)[1])
}
\arguments{
\item{mutnet.obj}{A dataframe of mutation info}

\item{base.ref.col}{Name of the column with the reference allele in dataframe}

\item{base.alt.col}{Name of the column with the alternative allele in dataframe}

\item{base.3p.col}{Name of the column with the base after the mutation in dataframe}

\item{base.5p.col}{Name of the column with the base before the mutation in dataframe}

\item{trinuc.pc}{Array of the proportion of each trinucleotide in the genome (for normalisation)}

\item{symetrize}{Boolean to symetrize reversable mutations or not}
}
\value{
A dataframe with each mutation and context, with frequencies and proportions
}
\description{
Compute frequencies and proportions of mutations with context
}

